Finished integration testing, minor changes to API.
Added code to allow for retriving labels by passing in the type rather than
having to use the specific method. This helps out mostly in testing but might be
useful for people who have mixed-type labels to worry about.

Added example.php to show exercising the whole system from end-to-end with the
testing endpoint. It will create a shipment, add two packages to the shipment
(using the fixture XML), retrieve the labels for both packages (and places them
in a ./temp/), and then closes the shipment.

The end result of example.php should be two new files in temp/ and a shipment in
Asendia's testing system that has been closed.